Skip ruby build if it's already done

+++ Rakefile
@@ -186,8 +186,10 @@ namespace :build do
 
     desc "Build #{build.name}"
     task build.name => ["#{build.name}-configure", "#{build.name}-libs", build.dest_dir] do
+      artifact = "rubies/ruby-#{build.name}.tar.gz"
+      next if File.exist?(artifact)
       sh "make install DESTDIR=#{build.dest_dir}", chdir: build.build_dir
-      sh "tar cfz rubies/ruby-#{build.name}.tar.gz -C rubies #{build.name}"
+      sh "tar cfz #{artifact} -C rubies #{build.name}"
     end
 
     task "#{build.name}-libs" => ["#{build.name}-configure"] do
